tests/wpt/tests/css/css-conditional/js/CSS-supports-L3.html

<!doctype html>
<meta charset="utf-8">
<title>CSS.supports() Level 3</title>
<link rel="help" href="https://www.w3.org/TR/css-conditional-3/#the-css-namespace">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  test(function() {
    assert_equals(CSS.supports("(color: red)"), true);
  }, "Single-argument form allows for declarations with enclosing parentheses");

  test(function() {
    assert_equals(CSS.supports("color: red"), true);
  }, "Single-argument form allows for declarations without enclosing parentheses");

  test(function() {
    assert_equals(CSS.supports("(color: red) and (color: blue)"), true);
  }, "Complex conditions allowed");

  test(function() {
    assert_equals(CSS.supports("not (foobar)"), true);
  }, "general_enclosed still parses");

  test(function() {
    assert_equals(CSS.supports("color: something-pointless var(--foo)"), true);
  }, "Variable references always parse");

  test(function() {
    assert_equals(CSS.supports("color: something-pointless(var(--foo))"), true);
  }, "Variable references in an unknown function always parse");

  test(function() {
    // no one-arg test for this as the with/without enclosing parentheses tests do this
    assert_equals(CSS.supports("color", "red"), true);
  }, "two argument form succeeds for known property");

  test(function() {
    assert_equals(CSS.supports("unknownproperty: blah"), false);
  }, "one argument form fails for unknown property");

  test(function() {
    assert_equals(CSS.supports("unknownproperty", "blah"), false);
  }, "two argument form fails for unknown property");

  test(function() {
    // https://github.com/w3c/csswg-drafts/issues/5929
    assert_equals(CSS.supports("unicode-range: U+0-7F"), false);
  }, "one argument form fails for unknown property (but known descriptor)");

  test(function() {
    // https://github.com/w3c/csswg-drafts/issues/5929
    assert_equals(CSS.supports("unicode-range", "U+0-7F"), false);
  }, "two argument form fails for unknown property (but known descriptor)");

  test(function() {
    // https://github.com/w3c/csswg-drafts/issues/5929
    assert_equals(CSS.supports("unicode-range: inherit"), false);
  }, "one argument form fails for unknown property (but known descriptor, universal value)");

  test(function() {
    // https://github.com/w3c/csswg-drafts/issues/5929
    assert_equals(CSS.supports("unicode-range", "inherit"), false);
  }, "two argument form fails for unknown property (but known descriptor, universal value)");

  test(function() {
    assert_equals(CSS.supports("width: blah"), false);
  }, "one argument form fails for invalid value");

  test(function() {
    assert_equals(CSS.supports("width", "blah"), false);
  }, "two argument form fails for invalid value");

  test(function() {
    assert_equals(CSS.supports("--foo: blah"), true);
  }, "one argument form succeeds for custom property");

  test(function() {
    assert_equals(CSS.supports("--foo", "blah"), true);
  }, "two argument form succeeds for custom property");
</script>
